requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
jinja2==3.1.4
itsdangerous==2.2.0
python-dotenv==1.0.1
cachetools==5.5.0

# AI & LangChain
langchain==0.2.6
langchain-core==0.2.10
langchain-community==0.2.6
langchain-groq==0.1.6
langchain-huggingface==0.0.3
langchain-text-splitters==0.2.2
langgraph==0.1.4

# PDF & Search
pdfplumber==0.11.2
faiss-cpu==1.8.0
duckduckgo-search==6.1.7

# Audio
gTTS==2.5.1
groq==0.9.0